Upgraded Django to 2.1
