+++ app/models/preset.py
@@ -12,7 +12,7 @@ logger = logging.getLogger('app.logger')
 class Preset(models.Model):
     owner = models.ForeignKey(User, blank=True, null=True, on_delete=models.CASCADE, help_text="The person who owns this preset")
     name = models.CharField(max_length=255, blank=False, null=False, help_text="A label used to describe the preset")
-    options = JSONField(default=list(), blank=True, help_text="Options that define this preset (same format as in a Task's options).",
+    options = JSONField(default=list, blank=True, help_text="Options that define this preset (same format as in a Task's options).",
                                validators=[validate_task_options])
     created_at = models.DateTimeField(default=timezone.now, help_text="Creation date")
     system = models.BooleanField(db_index=True, default=False, help_text="Whether this preset is available to every user in the system or just to its owner.")
