+++ nodeodm/models.py
@@ -40,7 +40,7 @@ class ProcessingNode(models.Model):
     api_version = models.CharField(max_length=32, null=True, help_text="API version used by the node")
     last_refreshed = models.DateTimeField(null=True, help_text="When was the information about this node last retrieved?")
     queue_count = models.PositiveIntegerField(default=0, help_text="Number of tasks currently being processed by this node (as reported by the node itself)")
-    available_options = fields.JSONField(default=dict(), help_text="Description of the options that can be used for processing")
+    available_options = fields.JSONField(default=dict, help_text="Description of the options that can be used for processing")
     token = models.CharField(max_length=1024, blank=True, default="", help_text="Token to use for authentication. If the node doesn't have authentication, you can leave this field blank.")
     max_images = models.PositiveIntegerField(help_text="Maximum number of images accepted by this node.", blank=True, null=True)
     odm_version = models.CharField(max_length=32, null=True, help_text="OpenDroneMap version used by the node")
@@ -221,12 +221,6 @@ class ProcessingNode(models.Model):
 
         from app.plugins import signals as plugin_signals
         plugin_signals.processing_node_removed.send_robust(sender=self.__class__, processing_node_id=pnode_id)
-
-
-    class Meta:
-        permissions = (
-            ('view_processingnode', 'Can view processing node'),
-        )
 
 
 # First time a processing node is created, automatically try to update
